refactor: types, helpers, and clean up root src
