+++ mcps/mcp-with-express/src/types/weather.types.ts
@@ -1,0 +1,33 @@
+export interface AlertFeature {
+  properties: {
+    event?: string;
+    areaDesc?: string;
+    severity?: string;
+    status?: string;
+    headline?: string;
+  };
+}
+export interface ForecastPeriod {
+  name?: string;
+  temperature?: number;
+  temperatureUnit?: string;
+  windSpeed?: string;
+  windDirection?: string;
+  shortForecast?: string;
+}
+
+export interface AlertsResponse {
+  features: AlertFeature[];
+}
+
+export interface PointsResponse {
+  properties: {
+    forecast?: string;
+  };
+}
+
+export interface ForecastResponse {
+  properties: {
+    periods: ForecastPeriod[];
+  };
+}
